chore(meta-service-test): move MetaNode tests to separate files
