+++ src/meta/service/tests/it/meta_node/mod.rs
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub(crate) mod meta_node_all;
-pub(crate) mod meta_node_expire;
 pub(crate) mod meta_node_kv_api;
+pub(crate) mod meta_node_kv_api_expire;
+pub(crate) mod meta_node_lifecycle;
+pub(crate) mod meta_node_replication;
+pub(crate) mod meta_node_request_forwarding;
+pub(crate) mod meta_node_seq_api;
